OGDF Snapshot 2017-02-16
Changes since last snapshot:
 * new graph generator customGraph() for quick generation of
   specific custom graphs
 * the graph size is no longer bounded by the stack size for the
   following algorithms:
   * biconnectedComponents()
   * strongComponents()
   * isForest()
 * new simple graph algorithm nodeDistribution() and
   degreeDistribution() as a special case
 * file format changes:
   * an STP reader for Graph (without edge weights and terminals)
     is now available, and it is included in the generic reader
   * the DMF reader and writer uses Graph and EdgeArray instead
     of EdgeWeightedGraph now
   * the DOT reader is now implemented non-recursively
     (which allows reading of large files while having low stack size)
   * Graph6, GDF, and TLP readers now clear the graph before reading
 * new minimum s-t-cut functionality:
   * MinSTCut module
   * MinSTCutMaxflow (replaces old MinSTCut)
   * MinSTCutBFS
   * MinSTCutDijkstra
 * new class DIntersectableRect replaces IntersectionRectangle
 * new PlanarSubgraphTree class as Maximum Planar Subgraph heuristic
 * bugfixes, code cleanup, improvement of code quality, C++11ifications,
   e.g., in:
   * OGDF initialization
   * the POOL_TS memory manager
   * GraphAttributes
   * GraphCopySimple
   * List, ListPure, SLIst, SListPure
   * DisjointSets
   * EdgeRouter
   * FMMMLayout
   * LCA
   * (fast_multipole_embedder::)ArrayGraph
   * (fast_multipole_embedder::)WSPD
   * the spring embedder force models
 * general (sometimes internal) changes:
   * removal of the doDestruction() function template
   * removal of a lot of shadowing of variables
   * fixes of one-definition-rule violations that led
     to weird problems
   * enum classes (scoped enums) are now used throughout OGDF
     instead of unscoped enums
   * typedefs are replaced by using for improved readability
     and better flexibility
   * declaration of several constructors as explicit to
     avoid implicit conversion where it is wrong
   * some auxiliary classes are moved into sub-namespaces and
     their files are often moved into sub-directories
   * slightly improved consistency in documentation
   * slightly improved code style consistency

Check the porting guide in doc/porting/snapshot.md for compatibility-breaking changes.

There are also CMake changes. The default target only compiles the OGDF
(and its dependencies). There are new targets "tests" and "examples"
doing the obvious, and a new target "build-all" that also compiles
tests and examples. Separate test targets are renamed from foo_bar_test to
test-foo-bar.

+++ include/ogdf/orthogonal/MinimumEdgeDistances.h
@@ -1,7 +1,7 @@
 /** \file
  * \brief Declaration of class MinimumEdgeDistances which maintains
  *        minimum distances between attached edges at a vertex
- *       (delta's and epsilon's)
+ *       (deltas and epsilons)
  *
  * \author Carsten Gutwenger
  *
@@ -39,11 +39,7 @@
 namespace ogdf {
 
 
-//---------------------------------------------------------
-// MinimumEdgeDistances
-// maintains input sizes for improvement compaction (delta's
-// and epsilon's)
-//---------------------------------------------------------
+//! Maintains input sizes for improvement compaction (deltas and epsilons)
 template <class ATYPE>
 class MinimumEdgeDistances {
 public:
@@ -59,7 +55,7 @@ public:
 		OGDF_ASSERT(int(s) <= 3);
 		OGDF_ASSERT(0 <= i);
 		OGDF_ASSERT(i <= 1);
-		return m_delta[v].info[s][i];
+		return m_delta[v].info[static_cast<int>(s)][i];
 	}
 
 	ATYPE &delta(node v, OrthoDir s, int i) {
@@ -67,7 +63,7 @@ public:
 		OGDF_ASSERT(int(s) <= 3);
 		OGDF_ASSERT(0 <= i);
 		OGDF_ASSERT(i <= 1);
-		return m_delta[v].info[s][i];
+		return m_delta[v].info[static_cast<int>(s)][i];
 	}
 
 	// returns epsilon_s(v)^i (with i = 0 => l, i = 1 => r)
@@ -76,7 +72,7 @@ public:
 		OGDF_ASSERT(int(s) <= 3);
 		OGDF_ASSERT(0 <= i);
 		OGDF_ASSERT(i <= 1);
-		return m_epsilon[v].info[s][i];
+		return m_epsilon[v].info[static_cast<int>(s)][i];
 	}
 
 	ATYPE &epsilon(node v, OrthoDir s, int i) {
@@ -84,7 +80,7 @@ public:
 		OGDF_ASSERT(int(s) <= 3);
 		OGDF_ASSERT(0 <= i);
 		OGDF_ASSERT(i <= 1);
-		return m_epsilon[v].info[s][i];
+		return m_epsilon[v].info[static_cast<int>(s)][i];
 	}
 
 	ATYPE separation() const {
